+++ airflow/models/param.py
@@ -19,15 +19,21 @@ from __future__ import annotations
 import contextlib
 import copy
 import json
+import logging
 import warnings
 from typing import TYPE_CHECKING, Any, ItemsView, MutableMapping, ValuesView
 
 from airflow.exceptions import AirflowException, ParamValidationError, RemovedInAirflow3Warning
 from airflow.utils.context import Context
+from airflow.utils.mixins import ResolveMixin
 from airflow.utils.types import NOTSET, ArgNotSet
 
 if TYPE_CHECKING:
     from airflow.models.dag import DAG
+    from airflow.models.dagrun import DagRun
+    from airflow.models.operator import Operator
+
+logger = logging.getLogger(__name__)
 
 
 class Param:
@@ -131,6 +137,16 @@ class ParamsDict(MutableMapping[str, Any]):
         self.__dict = params_dict
         self.suppress_exception = suppress_exception
 
+    def __bool__(self) -> bool:
+        return bool(self.__dict)
+
+    def __eq__(self, other: Any) -> bool:
+        if isinstance(other, ParamsDict):
+            return self.dump() == other.dump()
+        if isinstance(other, dict):
+            return self.dump() == other
+        return NotImplemented
+
     def __copy__(self) -> ParamsDict:
         return ParamsDict(self.__dict, self.suppress_exception)
 
@@ -216,7 +232,7 @@ class ParamsDict(MutableMapping[str, Any]):
         return resolved_dict
 
 
-class DagParam:
+class DagParam(ResolveMixin):
     """
     Class that represents a DAG run parameter & binds a simple Param object to a name within a DAG instance,
     so that it can be resolved during the run time via ``{{ context }}`` dictionary. The ideal use case of
@@ -252,3 +268,24 @@ class DagParam:
         with contextlib.suppress(KeyError):
             return context['params'][self._name]
         raise AirflowException(f'No value could be resolved for parameter {self._name}')
+
+
+def process_params(
+    dag: DAG,
+    task: Operator,
+    dag_run: DagRun | None,
+    *,
+    suppress_exception: bool,
+) -> dict[str, Any]:
+    """Merge, validate params, and convert them into a simple dict."""
+    from airflow.configuration import conf
+
+    params = ParamsDict(suppress_exception=suppress_exception)
+    with contextlib.suppress(AttributeError):
+        params.update(dag.params)
+    if task.params:
+        params.update(task.params)
+    if conf.getboolean('core', 'dag_run_conf_overrides_params') and dag_run and dag_run.conf:
+        logger.debug("Updating task params (%s) with DagRun.conf (%s)", params, dag_run.conf)
+        params.update(dag_run.conf)
+    return params.validate()